pyproject.toml
[project]
name = "play-sdl3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pysdl3>=0.9.8b9",
]
